reorganized some files on front-end for easier directory navigation and fixed rendering issues of app pages
