+++ client/src/pages/Dashboard.js
@@ -1,0 +1,19 @@
+
+import React, { useState } from "react";
+import Nav from "../components/Nav";
+import Footer from "../components/Footer";
+
+const Dashboard = () => {
+
+    return (
+        <div>
+            <header className="header">
+                <h1>YouFit</h1>
+            </header>
+            <h1>Dashboard Page!</h1> 
+            
+        </div>
+    );
+}
+
+export default Dashboard;
